cleaned up config and merged its modermode function into main file
